+++ test/element/click.js
@@ -6,18 +6,18 @@ const { sleep } = require('../helpers/')
 
 async function main() {
 
-   let chrome = await autoChrome({
+   const chrome = await autoChrome({
       executablePath: "C:/Program Files (x86)/Google/Chrome/Application/chrome.exe",
       userDataDir: userDataDir + 1,
       args: ['--start-maximized'],
       devtools: false,
    })
 
-   await chrome.page.goto('https://www.so.com/')
+   await chrome.page.goto('https://www.so.com/');
 
-   let nav = await chrome.page.$$('#header nav > a')
+   const nav = await chrome.page.$$('#header nav > a');
 
-   let [, , , element] = nav
+   const [, , , element] = nav;
 
    await element.click()