+++ test/input/touch.js
@@ -6,24 +6,24 @@ const { sleep } = require('../helpers/')
 
 async function main() {
 
-   let chrome = await autoChrome({
+   const chrome = await autoChrome({
       executablePath,
-      userDataDir,
+      userDataDir: userDataDir + 1,
       args: ['--start-maximized'],
       devtools: true,
    })
 
-   let page = chrome.page
+   const page = chrome.page
 
    await chrome.page.goto('D:/Nodejs/Project/auto-chrome/test/input/touch.html')
 
    await sleep(1000)
 
-   let input = await page.$('input')
+   const element = await page.$('input')
 
-   await input.click()
+   await element.click()
 
-   await input.type('hellow你好吗')
+   await element.type('hellow你好吗')
 
 }