+++ test/network.js
@@ -8,7 +8,7 @@ const { userAgent, viewport } = devices['iPhone 6']
 
 async function main() {
 
-   let chrome = await autoChrome({
+   const chrome = await autoChrome({
       executablePath: "C:/Program Files (x86)/Google/Chrome/Application/chrome.exe",
       userDataDir: userDataDir + 1,
       // devtools: true,
@@ -27,11 +27,11 @@ async function main() {
 
    await chrome.newPage('https://www.sogou.com')
 
-   let { page } = chrome
+   const { page } = chrome;
 
    await page.goto('https://m.so.com/')
 
-   let element = await page.click('#q')
+   const element = await page.click('#q')
 
    await element.type('ip')