+++ test/proxy_more.js
@@ -17,7 +17,7 @@ if (cluster.isMaster) {
 
    async function main() {
 
-      let chrome = await autoChrome({
+      const chrome = await autoChrome({
          executablePath: "C:/Program Files (x86)/Google/Chrome/Application/chrome.exe",
          userDataDir: userDataDir + 1,
          // devtools: true,